Revert "refactor(config): rewrite eslint presets in typescript"
This reverts commit 05cf311f820965f62a888934d886662f9f05e31c.
