--- packages/config/.prettierignore
@@ -1,2 +1,0 @@
-eslint/*.js
-eslint/*.d.ts
