+++ packages/config/eslint.config.js
@@ -1,3 +1,3 @@
-import { config } from './eslint/config.js'
+import { config } from './eslint/config.mjs'
 
 export default config
