--- packages/config/eslint/index.d.ts
@@ -1,1 +1,0 @@
-export { config, createTestConfig, createTypescriptConfig, ignoreConfig, testConfig, typescriptConfig } from './config.js';
