--- packages/config/eslint/index.ts
@@ -1,8 +1,0 @@
-export {
-  config,
-  createTestConfig,
-  createTypescriptConfig,
-  ignoreConfig,
-  testConfig,
-  typescriptConfig
-} from './config.js'
