+++ packages/config/package.json
@@ -27,7 +27,6 @@
     "globals": "^15.11.0"
   },
   "devDependencies": {
-    "@types/node": "^22.18.13",
     "@eslint/js": "^9.38.0",
     "@typescript-eslint/eslint-plugin": "^8.10.0",
     "@typescript-eslint/parser": "^8.10.0",
