--- packages/config/types/eslint-config-prettier/index.d.ts
@@ -1,4 +1,0 @@
-import type { Linter } from 'eslint'
-
-declare const config: Linter.Config
-export default config
