--- packages/config/types/globals/index.d.ts
@@ -1,4 +1,0 @@
-import type { Linter } from 'eslint'
-
-declare const globals: Record<string, Linter.Globals>
-export default globals
